ajout inscription / connexion / deconnexion

+++ Projet-JEE/src/main/java/com/example/Projet_JEE/entity/Evaluation_activite_Id.java
@@ -5,6 +5,6 @@ import jakarta.persistence.Embeddable;
 @Embeddable
 public class Evaluation_activite_Id {
 
-    private int idUtilisateur;
+    private Long idUtilisateur;
     private int idActivite;
 }
