+++ Projet-JEE/src/main/java/com/example/Projet_JEE/entity/Pathologie_utilisateur_Id.java
@@ -5,5 +5,5 @@ import jakarta.persistence.Embeddable;
 @Embeddable
 public class Pathologie_utilisateur_Id {
     private int idPathologie;
-    private int idUtilisateur;
+    private Long idUtilisateur;
 }
